docs: moved unit test for GetDriver to the respective eventsource and sensor directories
